+++ .ivy/raise-build-plugin-version.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+mvn --batch-mode versions:set-property versions:commit -Dproperty=project.build.plugin.version -DnewVersion=${2} -DallowSnapshots=true